requirements.txt
# Django and REST Framework
Django==6.0
djangorestframework==3.15.2
drf-spectacular==0.27.2

# Database
# SQLite3 is included with Python, no additional package needed
# PostgreSQL (for production deployment)
psycopg2-binary==2.9.9

# Production Server
gunicorn==21.2.0

# Production Utilities
whitenoise==6.6.0
dj-database-url==2.1.0

# Testing
coverage==7.6.1

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.2
